Edits to Data Structures I and II.
